Changes suggested by fjh review comments.
Estimated hours taken: 0.25
Branches: main

Changes suggested by fjh review comments.

README.MS-VisualC:
	Document TMPDIR better.

trace/mercury_trace_source.c:
	Change the loop counter from int to a long, just in case we
	one day target a 16bit system.

+++ trace/mercury_trace_source.c
@@ -351,7 +351,7 @@ MR_trace_source_attach(MR_Trace_Source_Server *server, int timeout,
         ** busy-wait for a few billion cycles, which should hopefully
         ** take a second or more.
         */
-        volatile int i;
+        volatile long i;
         for (i = 0; i < 100000000; i++)
         {}
 #endif
